Fix audit crash

+++ src/kotlin/io/dedyn/engineermantra/omega/bot/Auditing.kt
@@ -98,11 +98,14 @@ object Auditing {
                     guild.addRoleToMember(member, guild.getRoleById(1078829209616666705)!!).queue()
                 }
             }
-            for (role in botcGuild!!.getRoles()) {
+            for (role in botcGuild!!.roles) {
                 //Audit cleanup of old roles that need to be deleted. Hopefully can remove this later.
                 if(guild.getRolesByName(role.name, false).isEmpty())
                 {
-                    role.delete().queue()
+                    //Fix an audit crash related to us trying to delete a bot role that is in BOTC but not SC
+                    if(!role.isManaged) {
+                        role.delete().queue()
+                    }
                 }
             }
         }
